CSPM-Tool/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.10.0
python-multipart>=0.0.9

# ── Database ──────────────────────────────────────────────────────────────
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# ── Auth & Security ───────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# ── Scheduler ────────────────────────────────────────────────────────────
apscheduler>=3.10.0

# ── AWS ──────────────────────────────────────────────────────────────────
boto3>=1.34.0
botocore>=1.34.0

# ── Azure ────────────────────────────────────────────────────────────────
azure-identity>=1.16.0
azure-mgmt-resource>=23.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-compute>=31.0.0
azure-mgmt-network>=25.0.0
azure-mgmt-keyvault>=10.3.0
azure-mgmt-sql>=3.0.1
azure-mgmt-monitor>=6.0.0
azure-mgmt-web>=7.0.0
azure-mgmt-containerservice>=28.0.0
azure-mgmt-containerregistry>=10.0.0
azure-mgmt-cosmosdb>=9.0.0
azure-mgmt-redis>=14.0.0

# ── Utilities ─────────────────────────────────────────────────────────────
PyYAML>=6.0.1
python-dotenv>=1.0.0
